Move hacks to hacks package
